Add files via upload
# File_Packer_Unpacker

```This Application Is Based On Zipping And Unzipping The Directory```

### PLATFORM REQUIRED :   
```Windows Platform Or Linux Distributions.```
### ARCHITECTURAL REQUIREMENT :  
```INTEL 32 Bit Processor Or Higher.```
### USER INTERFACE :             
```Command Line Interface And GUI```
### TECHNOLOGY USED : 
```Java Programming.```

# About This Application :

```1) This Application Is Used To Perform Packing And Unpacking Activity For Multiple Types Of Files. ```

```2) In Case Of Packing Activity I Maintain One File Which Contains Metadata And Data Of Multiple Files From Specified Directory. ```

```3) In Case Of Unpacking Activity I Extract All Data From Packed Files And According To Its Metadata It Create All Files. ```

```4) In This Application I Use Java As Frontend As Well As Backend Also For Platform Independency. ```